+++ libs/fsm/tests/event_queue.c
@@ -125,7 +125,7 @@ static enum am_rc fsmq_a(struct am_fsm* fsm, const struct am_event* event) {
     case AM_EVT_A: {
         me->log("a-A;");
         const struct am_event* e =
-            am_event_allocate(me->alloc, /*id=*/AM_EVT_B, sizeof(*e));
+            am_event_allocate(me->alloc, /*event_id=*/AM_EVT_B, sizeof(*e));
         am_event_queue_push_back(&me->event_queue, e);
         return am_fsm_tran(fsm, fsmq_b);
     }