+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [0-based versioning](https://0ver.org/).
 ### Changed
 
 - Allocate `struct am_event_async_hub` on the caller side
+- Use `uint16_t` for event IDs in event API
 
 ### Fixed